Support multiple completion processes
This is required for Tramp support.

Use new function  `fsharp-ac-completion-process' to get completion
process for a specified file. The new functions
`fsharp-ac-completion-process-add' and
`fsharp-ac-completion-process-del' are used to add/remove new processes.

Also remove of global fsharp `fsharp-ac--completion-bufname': Unique
buffer names are created. There is no need to store the buffer names,
because they are directly associated with a process.

Refs #57.