Fix FSAC hanging issue
The protocol chatter sent to FSAC to parse some code goes like this:

parse "filename"
...F# code...
...F# code...
...F# code...
<<EOF>>

<<EOF>> is not an actual EOF, but literally the string <<EOF>>, which must be
preceded by a newline and followed by a newline.

The issue was that typing at the end of the buffer would cause the buffer's
contents to be sent to FSAC, but because the last byte of the buffer would not
be the conventional appended-at-save newline, FSAC would be sent this instead:

parse "filename"
...F# code...
...F# code...
...F# code...<<EOF>

...and thus continue waiting for input indefinitely, causing things like type
info and Flycheck to stop working.

Fixes #119.