Change directory to same as build script for compile command
Build scripts that use relative paths can get confused if they're called
from a different directory, so first cd to the same directory as the
script, and then run it.

Also add quotes around the paths when running the command to avoid
issues with some characters. I used this instead of shell-quote-command
because that was also escaping the : character in file paths on windows,
which made the commands invalid.