Fix a small bug in font locking for active records
This commit fixes a bug in which only the final name in a multi-symbol active
record would be font locked. The trouble is that the regexp requires the closing
banana clip |) as part of the pattern; the final paren should be _optional_.
Thus, the only meaningful change is adding a ? inside the regexp. A newline is
introduced to make the whole thing slightly more readable.