Update attribute font locking
Previously, attributes like `[<EntryPoint>]` would lock just fine, while
attributes like `[<Options("arg")>]` would fail to lock at all. This was because
the regex couldn't handle the parens and inner string, so would skip the whole
match.

This commit fixes that, making sure the parens lock in the same face as the rest
of the attribute, while any provided inner properties lock as a string.