Tell comint to set TERM=emacs for fsi, and don't force using a pipe
We want to be able to use fsi's tab-completion, but we can't if we have
to pass --readline- as it turns off readline support. comint reports itself
as a dumb-terminal by default, which causes fsi to treat it like one. Setting
TERM=emacs (or likely and value that's not 'dumb') appears to let us interact
with fsi in the way we want to.

The previous combination of (process-connection-nil) and setting --readline- was
probably done to get around fsi's behavior when using a pty and TERM=dumb, but Im
not 100% sure of that.

This setup for init-ing the inferior fsharp process might need revisted in the future.