Refactoring Pass: Extract Method from `fsharp-compute-indentation`
`fsharp-compute-indentation` is one of the critical driving methods of this
mode's indentation system. At the start of this commit, it was also a very long,
very complex method. It needs better documentation and it needs to be under
test. Therefore: extract method.

This commit pulls a variety of method out of `fsharp-compute-indentation`,
seeking to preserve basically identical functionality while capturing pieces of
logic in smaller, better documented pieces. Functions are currently named after
my best read of what they _do_, but that's hard to be confident about.

The _next_ piece of work will be getting these functions under test, which will
allow me to pin down and describe their functionality (and its limits!).