This is A Lot, but it is, broadly, better.
Okay. This got really messy, and I need to pin everything down.

It turns out that F# indentation is _really hard_. Everything is relative, and
the only thing that matters is that *blocks line up*. That's fine, but also
hard. There are many cases where the "normal" behavior is to line up with a
particular part of the preceding line, rather than with a particular indent
level.

So.

This is slowly taking the approach that we should maintain a list of the levels
of indent, and we should see:

1. Ret *mostly* goes to the same level, except in certain cases when we
might *indent*
2. Tab will first increase indent one level, then go back to start, then
decrease through all the preceeding levels.
3. Indentation is for semantics, lining up blocks correctly so the compiler can
interpret them correctly. *Formatting* should be handled by Fantomas, either via
LSP or by calling it directly.