Fix interference with company-mode
Company-mode has deprecated the variables company-auto-complete and
company-auto-complete-chars in favor of company-auto-commit and
company-auto-commit-chars for the same purposes.
The references to the old variable names are kept als deprecated aliases.
Creating local variables with the deprecated variable names seems to cause
issues when activating company-mode which complains with "Don't know how to make
a localized variable an alias".
This seems to prevent reliable auto-completion.