Make `eglot-fsharp' tramp aware
Install "fsautocomplete" on the remote host if `default-directory' is
a tramp filename.

Add pipe-workaround for remote fsautocomplete processes: There seems
to be a problem related to CRLF coded input in remote fsautocomplete
processes. Wrapping the process using a cat pipe fixes the
issue.

Even when setting `process-connection-type' to nil or 'pipe there is
still a pseudo tty for the started remote process:
https://www.gnu.org/software/emacs/manual/html_node/tramp/Remote-processes.html#Remote-process-connection-type

Fixes #324